Add a separate topic short name (v2)
And update the toolkit. The toolkit will now prefer the short name
for the document title, show the activity name first and only
show the topic name for the "welcome" activity.

The initial motivation for this was that all tabs of all apps started
with something like "Electronic ..." and weren't distinguishable.