Don't allow submissions in case the subject or reference number are missing
In case they are missing or empty we show an error message to the user
and don't allow submitting the order.

Since there are now 4 things that are required to be filled out before
a user can submit, help the user by creating a separate error message for
each missing data entry.

This also merges the validation code paths for both the list and the single
entry view into one function while at it.