Set a default refrence number when creating a request
Sets a default reference number when creating a dispatch request,
sourced from the translations, which can be overridden via translation
overrides.

In case a PDF with a reference number gets added after the request is
created (which is the default flow) we only set the reference number
in case it isn't set, is empty, or matches the default.

The later is a bit hacky, since a valid number could match the default,
but this is unlikely to happen, and otherwise we'd need to keep track
if the default was set in the backend.