build: add the new handy-signatur host to the CSP header
When using pdf-as in the backend the iframe will show the host specified
by mobile.sign.url in the backend. We are switching the backend to a
new endpoint, so we have to adjust the CSP to allow loading from that new host.

Keep www.handy-signatur.at for backwards compat.